Answer_Selection/code/requirements.txt
#python==3.6.6

backcall==0.1.0
chardet==3.0.4
decorator==4.4.0
einops==0.1.0
fire==0.1.3
greenlet==0.4.15
idna==2.8
inflection==0.3.1
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
msgpack==0.6.1
munch==2.3.2
neovim==0.3.1
nltk==3.4.1
numpy==1.16.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.0
pynvim==0.3.2
pytorch-ignite==0.2.0
requests==2.22.0
six==1.12.0
tensorboard==1.14.0
tensorboardX==1.7
tensorflow-gpu==1.14.0
torch>=1.1.0
torchtext>=0.3.1
torchvision>=0.2.2.post3
tqdm>=4.32.1
traitlets==4.3.2
urllib3==1.25.2
wcwidth==0.1.7
rouge==0.3.2
unidecode==1.1.1
